backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coulisses-crew-api"
version = "0.6.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.8.2",
    "httpx==0.27.0",
    "pyjwt==2.9.0",
    "SQLAlchemy==2.0.32",
    "passlib[bcrypt]==1.7.4",
    "alembic==1.13.2",
    "psycopg[binary]==3.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "pytest-cov==5.0.0",
    "ruff==0.6.8",
    "mypy==1.11.2",
    "types-requests",
    "requests==2.32.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
disallow_untyped_defs = false
